% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abetadms__datatable_to_matrix.R
\name{abetadms__datatable_to_matrix}
\alias{abetadms__datatable_to_matrix}
\title{abetadms__datatable_to_matrix}
\usage{
abetadms__datatable_to_matrix(input_dt, variable_name = "fitness")
}
\arguments{
\item{input_dt}{input data.table (required)}

\item{variable_name}{name of variable to use for heatmap cells (defaut:"fitness")}
}
\value{
a matrix for heamtmap plotting
}
\description{
Convert data.table to matrix for heatmap plotting (single AA mutants only).
}
