% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abetadms__glm_prenorm.R
\name{abetadms__glm_prenorm}
\alias{abetadms__glm_prenorm}
\title{abetadms__glm_prenorm}
\usage{
abetadms__glm_prenorm(x)
}
\arguments{
\item{x}{data.frame to be normalised (required)}
}
\value{
A normalised data.frame
}
\description{
Normalise data.frame columns.
}
