% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abetadms__load_fitness.R
\name{abetadms__load_fitness}
\alias{abetadms__load_fitness}
\title{abetadms__load_fitness}
\usage{
abetadms__load_fitness(dms_dir, object = F, read_filter = T)
}
\arguments{
\item{dms_dir}{path to variant fitness directory (required)}

\item{object}{load from R data object (default:F)}

\item{read_filter}{filter reads based on is.reads0 (default:T)}
}
\value{
a variant fitness data.table
}
\description{
Load variant fitness estimates.
}
