% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abetadms__plot_fitness_distributions.R
\name{abetadms__plot_fitness_distributions}
\alias{abetadms__plot_fitness_distributions}
\title{abetadms__plot_fitness_distributions}
\usage{
abetadms__plot_fitness_distributions(input_dt, output_file,
  boundary_list = NULL, boundary_colour_list = NULL, xlim = NULL,
  colour_scheme = NULL, width = 6, height = 13,
  facet_by_mut_type = T)
}
\arguments{
\item{input_dt}{a data.table with columns: STOP, fitness_cond, region, mut_type (required)}

\item{output_file}{path to output plot (required)}

\item{boundary_list}{a list of fitness values to highlight with vertical lines on plot}

\item{boundary_colour_list}{a list of colours corresponding the boundary_list vertical lines}

\item{xlim}{x axis limits}

\item{colour_scheme}{colour scheme list}

\item{width}{plot width}

\item{height}{plot height}

\item{facet_by_mut_type}{facet by mut_type (default:T)}
}
\value{
Nothing
}
\description{
Plot fitness distributions (and optional vertical lines highlighting specific fitness values).
}
