% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abetadms__run_bayesian_double_fitness.R
\name{abetadms__run_bayesian_double_fitness}
\alias{abetadms__run_bayesian_double_fitness}
\title{abetadms__run_bayesian_double_fitness}
\usage{
abetadms__run_bayesian_double_fitness(wt_dt, singles_dt, doubles_dt,
  outpath, all_reps, min_mean_input_read_count = 10,
  min_input_read_count_doubles = 50, lam_d = 0.025, numCores = 10)
}
\arguments{
\item{wt_dt}{data.table with WT (required)}

\item{singles_dt}{data.table with double AA mutants (required)}

\item{doubles_dt}{data.table with double AA mutants (required)}

\item{outpath}{output path for plots and saved objects (required)}

\item{all_reps}{list of replicates to retain (required)}

\item{min_mean_input_read_count}{minimum mean input read count for high confidence variants (default:10)}

\item{min_input_read_count_doubles}{minimum input read count for doubles used to derive prior for Bayesian doubles correction (default:50)}

\item{lam_d}{Poisson distribution for score likelihood (default:0.025)}

\item{numCores}{Number of available CPU cores (default:1)}
}
\value{
Nothing
}
\description{
Estimate fitness of double mutants using bayesian framework.
}
