% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abetadms_AB42_structure_propensities.R
\name{abetadms_AB42_structure_propensities}
\alias{abetadms_AB42_structure_propensities}
\title{abetadms_AB42_structure_propensities}
\usage{
abetadms_AB42_structure_propensities(result_dir, outpath, miscpath,
  DMS2structure_path, colour_scheme, execute = TRUE,
  rerun_structure = F)
}
\arguments{
\item{result_dir}{directory with AB42 structure match results (required)}

\item{outpath}{output path for plots and saved objects (required)}

\item{miscpath}{path to misc scripts and data directory (required)}

\item{DMS2structure_path}{Path to DMS2structure repository (required)}

\item{colour_scheme}{colour scheme file (required)}

\item{execute}{whether or not to execute the analysis (default: TRUE)}

\item{rerun_structure}{re-run structure analysis? (default:F)}
}
\value{
Nothing
}
\description{
AB42 structure predictions.
}
