% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abetadms_agg_tools_mutant_effects.R
\name{abetadms_agg_tools_mutant_effects}
\alias{abetadms_agg_tools_mutant_effects}
\title{abetadms_agg_tools_mutant_effects}
\usage{
abetadms_agg_tools_mutant_effects(fitness_dt, outpath,
  aggtool_results_file, aggscale_file, colour_scheme, execute = TRUE)
}
\arguments{
\item{fitness_dt}{data.table with single and double mutant codes (required)}

\item{outpath}{output path for plots and saved objects (required)}

\item{aggtool_results_file}{path to aggregation tool results file (required)}

\item{aggscale_file}{path to aggregation scale file (required)}

\item{colour_scheme}{colour scheme file (required)}

\item{execute}{whether or not to execute the analysis (default: TRUE)}
}
\value{
A data.table with single and double mutant effects from aggregation tool predictions
}
\description{
Calculate single and double mutant effects from aggregation tool predictions (using single mutants).
}
