% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abetadms_combine_fitness.R
\name{abetadms_combine_fitness}
\alias{abetadms_combine_fitness}
\title{abetadms_combine_fitness}
\usage{
abetadms_combine_fitness(fitness_list, disease_mut_file, outpath,
  colour_scheme, execute = TRUE)
}
\arguments{
\item{fitness_list}{named list of folder paths with fitness estimates (required)}

\item{disease_mut_file}{table of human disease mutations and classifications (required)}

\item{outpath}{output path for plots and saved objects (required)}

\item{colour_scheme}{colour scheme file (required)}

\item{execute}{whether or not to execute the analysis (default: TRUE)}
}
\value{
A data.table with normalised fitness data
}
\description{
Combine fitness estimates from two different DMS experiments (regions).
}
