% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abetadms_epistasis_analysis.R
\name{abetadms_epistasis_analysis}
\alias{abetadms_epistasis_analysis}
\title{abetadms_epistasis_analysis}
\usage{
abetadms_epistasis_analysis(fitness_path, miscpath, DMS2structure_path,
  numCores = 1, execute = TRUE)
}
\arguments{
\item{fitness_path}{path to R data object with single and double mutant fitness values (required)}

\item{miscpath}{path to misc scripts and data directory (required)}

\item{DMS2structure_path}{Path to DMS2structure repository (required)}

\item{numCores}{Number of available CPU cores (default:1)}

\item{execute}{whether or not to execute the analysis (default: TRUE)}
}
\value{
Nothing
}
\description{
Epistasis analysis.
}
