% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abetadms_fitness_model_summary.R
\name{abetadms_fitness_model_summary}
\alias{abetadms_fitness_model_summary}
\title{abetadms_fitness_model_summary}
\usage{
abetadms_fitness_model_summary(fitness_dt, outpath, colour_scheme,
  execute = TRUE)
}
\arguments{
\item{fitness_dt}{data.table with single and double mutant fitness values and their mutant effects from AA PCA (required)}

\item{outpath}{output path for plots and saved objects (required)}

\item{colour_scheme}{colour scheme file (required)}

\item{execute}{whether or not to execute the analysis (default: TRUE)}
}
\value{
Nothing
}
\description{
Dot plots showing explained variance of models to predict variant fitness.
}
