% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abetadms_PWI_heatmaps.R
\name{abetadms_PWI_heatmaps}
\alias{abetadms_PWI_heatmaps}
\title{abetadms_PWI_heatmaps}
\usage{
abetadms_PWI_heatmaps(PWI_dir_list, outpath, colour_scheme,
  execute = TRUE)
}
\arguments{
\item{PWI_dir_list}{list of PWI directories (required)}

\item{outpath}{output path for plots and saved objects (required)}

\item{colour_scheme}{colour scheme file (required)}

\item{execute}{whether or not to execute the analysis (default: TRUE)}
}
\value{
Nothing
}
\description{
Plot pair-wise interacton matrix heatmaps.
}
