% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abetadms_single_mutant_heatmaps.R
\name{abetadms_single_mutant_heatmaps}
\alias{abetadms_single_mutant_heatmaps}
\title{abetadms_single_mutant_heatmaps}
\usage{
abetadms_single_mutant_heatmaps(fitness_dt, disease_mut_file, outpath,
  colour_scheme, execute = TRUE)
}
\arguments{
\item{fitness_dt}{data.table with single mutant fitness values (required)}

\item{disease_mut_file}{table of human disease mutations and classifications (required)}

\item{outpath}{output path for plots and saved objects (required)}

\item{colour_scheme}{colour scheme file (required)}

\item{execute}{whether or not to execute the analysis (default: TRUE)}
}
\value{
Nothing
}
\description{
Combine fitness estimates from two different DMS experiments (regions).
}
