% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abetadms_wt_helix_propensity.R
\name{abetadms_wt_helix_propensity}
\alias{abetadms_wt_helix_propensity}
\title{abetadms_wt_helix_propensity}
\usage{
abetadms_wt_helix_propensity(fitness_dt, outpath, aaprop_file,
  aaprop_file_selected, wtfasta_path, colour_scheme, execute = TRUE)
}
\arguments{
\item{fitness_dt}{data.table with single and double mutant fitness values and their mutant effects from AA PCA (required)}

\item{outpath}{output path for plots and saved objects (required)}

\item{aaprop_file}{path to amino acid properties file (required)}

\item{aaprop_file_selected}{path to file with selected subset of identifiers}

\item{wtfasta_path}{path to WT FASTA sequence file (required)}

\item{colour_scheme}{colour scheme file (required)}

\item{execute}{whether or not to execute the analysis (default: TRUE)}
}
\value{
Nothing
}
\description{
Hydrophobicity of WT and fitness hotspot line plot.
}
